test(csharp): add CloudFetch failure injection tests
Added comprehensive CloudFetch failure scenarios:
- cloudfetch_expired_link: Link expiration with fallback to FetchResults
- cloudfetch_azure_403: Azure 403 Forbidden error handling
- cloudfetch_timeout: 65s timeout scenario
- cloudfetch_connection_reset: Connection reset during download
- Normal CloudFetch: Baseline test without failures

Removed SessionLifecycleTests.cs as it requires Thrift operation parsing.

All 5 CloudFetch tests passing. Proxy detects CloudFetch via cloud storage
domain detection (blob.core.windows.net, s3.amazonaws.com, storage.googleapis.com).
